requirements.txt
torch==1.12.1
transformers==4.21.1
sympy==1.10.1
PyPDF2==1.26.0
pandas==1.4.2
numpy==1.22.4
scikit-learn==1.1.1
jupyter==1.0.0
matplotlib==3.5.2
seaborn==0.11.2
aspose-pdf==25.2.0